% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ofv_ls.R
\name{calc_ofv_ls}
\alias{calc_ofv_ls}
\title{Calculate objective function value for LS fit}
\usage{
calc_ofv_ls(dv, ipred, res_sd, weights = 1, ...)
}
\arguments{
\item{dv}{observed data points}

\item{ipred}{individual predictions}

\item{res_sd}{standard deviation of the observations}

\item{weights}{weights vector for the observed data}

\item{...}{dummy parameters, not used}
}
\description{
Calculate objective function value for LS fit
}
