% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ofv_map.R
\name{calc_ofv_map}
\alias{calc_ofv_map}
\title{Calculate objective function value for MAP Bayesian fit}
\usage{
calc_ofv_map(
  eta,
  omega,
  omega_inv,
  omega_eigen,
  dv,
  ipred,
  res_sd,
  weights = 1,
  weight_prior = 1,
  include_omega = TRUE,
  include_error = TRUE
)
}
\arguments{
\item{eta}{eta's}

\item{omega}{full omega matrix}

\item{omega_inv}{inverse of omega matrix. Passed to this function to avoid doing computations in each iteration of the search.}

\item{omega_eigen}{eigenvalue decomposation (logged) of omega matrix. Passed to this function to avoid doing computations in each iteration of the search.}

\item{dv}{observed data points}

\item{ipred}{individual predictions}

\item{res_sd}{standard deviation of the observations}

\item{weights}{weights vector for the observed data}

\item{weight_prior}{relative weight of the population priors}

\item{include_omega}{Include deviation from population parameters in the likelihood?}

\item{include_error}{Include residual error in the likelihood?}
}
\description{
Calculate objective function value for MAP Bayesian fit
}
