% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ofv_map.R
\name{calc_ofv_map_slow}
\alias{calc_ofv_map_slow}
\title{Calculate objective function value for MAP Bayesian fit
This implementation is slower, but leaving available for use
in case of stability issues}
\usage{
calc_ofv_map_slow(
  eta,
  omega,
  dv,
  ipred,
  res_sd,
  weights = 1,
  weight_prior = 1,
  include_omega = TRUE,
  include_error = TRUE
)
}
\arguments{
\item{eta}{eta's}

\item{omega}{full omega matrix}

\item{dv}{observed data points}

\item{ipred}{individual predictions}

\item{res_sd}{standard deviation of the observations}

\item{weights}{weights vector for the observed data}

\item{weight_prior}{relative weight of the population priors}

\item{include_omega}{Include deviation from population parameters in the likelihood?}

\item{include_error}{Include residual error in the likelihood?}
}
\value{
Returns a vector of probabilities based on the probability of parameter 
        estimates given their multivariate distribution and the prediction errors
        based on these parameters. The product of the returned probabilities 
        is the likelihood.
}
\description{
Calculate objective function value for MAP Bayesian fit
This implementation is slower, but leaving available for use
in case of stability issues
}
