% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block_from_cv.R
\name{create_block_from_cv}
\alias{create_block_from_cv}
\title{Create an variance-covariance matrix from specified CV
Mostly useful for creating IOV blocks}
\usage{
create_block_from_cv(cv = 0.1, n = 4)
}
\arguments{
\item{cv}{CV of}

\item{n}{number of occassions to allow}
}
\description{
Create an variance-covariance matrix from specified CV
Mostly useful for creating IOV blocks
}
