% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res_from_prob.R
\name{dnorminv}
\alias{dnorminv}
\title{Inverse of the dnorm function in R.}
\usage{
dnorminv(y)
}
\arguments{
\item{y}{density}
}
\description{
This is not the same as qnorm, which is the inverse of pnorm (the CDF).
}
\details{
Taken from: https://stackoverflow.com/questions/19589191/the-reverse-inverse-of-the-normal-distribution-function-in-r
}
