% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_to_triangle.R
\name{full_to_triangle}
\alias{full_to_triangle}
\title{Convert full matrix to lower/upper triangle (vector)}
\usage{
full_to_triangle(d)
}
\arguments{
\item{d}{matrix}
}
\description{
Convert full matrix to lower/upper triangle (vector)
}
