% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood_of_data.R
\name{get_likelihood_of_data}
\alias{get_likelihood_of_data}
\title{Get the likelihood of observed data given a specific error function}
\usage{
get_likelihood_of_data(data, ipred, error, weights = NULL)
}
\arguments{
\item{data}{vector of observed data points}

\item{ipred}{vector of predicted values,}

\item{error}{list specifying error model (list of `prop` and `add`), assuming normal distribution}

\item{weights}{vector of weights for observations (not used by default, `NULL`)}
}
\description{
Get the likelihood of observed data given a specific error function
}
