% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mahalanobis.R
\name{get_mahalanobis}
\alias{get_mahalanobis}
\title{Calculates the Mahalanobis distance from observations, predictions and residual error-weighted predictions}
\usage{
get_mahalanobis(y, ipred, w_ipred, ltbs = FALSE)
}
\arguments{
\item{y}{observed values (DV)}

\item{ipred}{individualized predicted values}

\item{w_ipred}{weighted individualized predicted values}

\item{ltbs}{logical indicating whether both sides should be log transformed}
}
\description{
Calculates the Mahalanobis distance from observations, predictions and residual error-weighted predictions
}
