% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_np_estimates.R
\name{get_np_estimates}
\alias{get_np_estimates}
\title{Get non-parametric estimates given a set of support points, model, and data}
\usage{
get_np_estimates(
  parameter_grid = NULL,
  error = list(prop = 0.1, add = 0.1),
  model = NULL,
  covariates = NULL,
  regimen = NULL,
  data = NULL,
  t_obs = NULL,
  weights = NULL,
  ...
)
}
\arguments{
\item{parameter_grid}{grid (data.frame) of support points}

\item{error}{error model}

\item{model}{`PKPDsim` model}

\item{covariates}{`PKPDsim` covariates object}

\item{regimen}{`PKPDsim` regimen}

\item{data}{vector of observed data}

\item{t_obs}{vector of observation times, matching `data`}

\item{weights}{vector of weights passed to `get_map_estimates()`}

\item{...}{passed to `get_map_estimates()`}
}
\description{
Get non-parametric estimates given a set of support points, model, and data
}
