% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradient_weights.R
\name{gradient_weights}
\alias{gradient_weights}
\title{Gradient weighting}
\usage{
gradient_weights(time = NULL, t_full_weight = 0)
}
\arguments{
\item{time}{vector of time}

\item{t_full_weight}{time after which weight should be 1}
}
\description{
Weighting increases from 0 to 1 by a linear gradient. At `t_full_weight` the weighting is 1.
}

