% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_blocks.R
\name{join_blocks}
\alias{join_blocks}
\title{Join omega blocks}
\usage{
join_blocks(om1, om2, as_triangle = TRUE)
}
\arguments{
\item{om1}{block 1}

\item{om2}{block 2}

\item{as_triangle}{should output block be returned as triangle (vector) or not (matrix)?}
}
\description{
Join omega blocks
}
