% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_func_generic.R
\name{ll_func_generic}
\alias{ll_func_generic}
\title{Generic likelihood function}
\usage{
ll_func_generic(
  data,
  eta01,
  eta02,
  eta03,
  eta04,
  eta05,
  eta06,
  eta07,
  eta08,
  eta09,
  eta10,
  eta11,
  eta12,
  eta13,
  eta14,
  eta15,
  eta16,
  eta17,
  eta18,
  eta19,
  eta20,
  eta21,
  eta22,
  eta23,
  eta24,
  parameters,
  covariates = NULL,
  regimen = regimen,
  lagtime = NULL,
  omega_full = omega_full,
  error = error,
  model,
  sig,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{eta01}{eta1}

\item{eta02}{eta2}

\item{eta03}{eta3}

\item{eta04}{eta4}

\item{eta05}{eta5}

\item{eta06}{eta6}

\item{eta07}{eta7}

\item{eta08}{eta8}

\item{eta09}{eta9}

\item{eta10}{eta10}

\item{eta11}{eta11}

\item{eta12}{eta12}

\item{eta13}{eta13}

\item{eta14}{eta14}

\item{eta15}{eta15}

\item{eta16}{eta16}

\item{eta17}{eta17}

\item{eta18}{eta19}

\item{eta19}{eta19}

\item{eta20}{eta20}

\item{eta21}{eta21}

\item{eta22}{eta22}

\item{eta23}{eta23}

\item{eta24}{eta24}

\item{parameters}{list of parameters}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{regimen}{regimen}

\item{lagtime}{vector of lagtimes for each compartment in the model, either
numeric vector, or vector of parameters.}

\item{omega_full}{full omega matrix}

\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}

\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{sig}{signficance, used in optimization}

\item{verbose}{show more output}

\item{...}{passed on to PKPDsim}
}
\description{
Generic likelihood function
}
