% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\docType{data}
\name{log2pi}
\alias{log2pi}
\title{Constant used in some computations}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
log2pi
}
\description{
Constant used in some computations
}
\keyword{datasets}
