% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_triangle_mat_size.R
\name{lower_triangle_mat_size}
\alias{lower_triangle_mat_size}
\title{Get the size of matrix specified as vector elements of the lower triangle}
\usage{
lower_triangle_mat_size(mat)
}
\arguments{
\item{mat}{vector of matrix elements of lower triangle}
}
\description{
Get the size of matrix specified as vector elements of the lower triangle
}
