% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_wrapper.R
\name{mle_wrapper}
\alias{mle_wrapper}
\title{Wrapper function around optimization function (e.g. `optim`)}
\usage{
mle_wrapper(
  minuslogl,
  start,
  method = "BFGS",
  optimizer = c("optim"),
  data = NULL,
  control = list(reltol = 1e-05),
  skip_hessian = FALSE
)
}
\arguments{
\item{minuslogl}{function that calculates -log(L) based on parameters and other additional `data` passed to it.}

\item{start}{initial estimates for the parameters}

\item{method}{estimation method passed to optimization library.}

\item{optimizer}{optimizer used. Currently only `optim` is supported.}

\item{data}{additional data required for `minuslogl` function to perform optimization.}

\item{control}{additional parameters that control the optimization process.}

\item{skip_hessian}{skip the calculation of the Hessian and variance-covariance matrix at the MAP estimates?}
}
\description{
Originally the function used for this was bbmle::mle(), but that
function was needlessly complex and had several outstanding
development issues.
}
