% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_fit_wrapper.R
\name{np_fit_wrapper}
\alias{np_fit_wrapper}
\title{Wrapper for non-parametric fit}
\usage{
np_fit_wrapper(
  obj,
  model,
  regimen,
  data,
  covariates = NULL,
  weights = 1,
  np_settings = list()
)
}
\arguments{
\item{obj}{fit object}

\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{regimen}{regimen}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Note that `weights` argument will also affect residuals (residuals will be 
scaled too).}

\item{np_settings}{list with settings for non-parametric estimation (if 
selected), containing any of the following: `error`, `grid_span`, grid_size`,
 `grid_exponential`}
}
\description{
Wrapper for non-parametric fit
}
