% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_settings_default.R
\docType{data}
\name{np_settings_default}
\alias{np_settings_default}
\title{Default settings for non-parametric fit}
\format{An object of class \code{list} of length 8.}
\usage{
np_settings_default
}
\description{
Default settings for non-parametric fit
}
\keyword{datasets}
