% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_censoring.R
\name{parse_censoring}
\alias{parse_censoring}
\title{Check if censoring functionality needs to be used, 
and create index for censored data}
\usage{
parse_censoring(censoring, data, verbose = FALSE)
}
\arguments{
\item{censoring}{label for column specifying censoring. If value in dataset 
in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{verbose}{show more output}
}
\description{
Check if censoring functionality needs to be used, 
and create index for censored data
}
