% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_error.R
\name{parse_error}
\alias{parse_error}
\title{Parse specified error magnitudes}
\usage{
parse_error(error)
}
\arguments{
\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}
}
\description{
Parse specified error magnitudes
}
