% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fixed.R
\name{parse_fixed}
\alias{parse_fixed}
\title{Parse fixed parameters}
\usage{
parse_fixed(fixed, parameters)
}
\arguments{
\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{parameters}{list of parameters}
}
\description{
Parse fixed parameters
}
