% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_omega_matrix.R
\name{parse_omega_matrix}
\alias{parse_omega_matrix}
\title{Parse omega matrix into full matrix, and
perform some checks}
\usage{
parse_omega_matrix(omega, parameters, fixed)
}
\arguments{
\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{parameters}{list of parameters}

\item{fixed}{fix a specific parameters, supplied as vector of strings}
}
\description{
Parse omega matrix into full matrix, and
perform some checks
}
