% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_t_obs.R
\name{parse_t_obs}
\alias{parse_t_obs}
\title{Parse observation times}
\usage{
parse_t_obs(data)
}
\arguments{
\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}
}
\description{
Parse observation times
}
