% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_weight_prior.R
\name{parse_weight_prior}
\alias{parse_weight_prior}
\title{Parse the weight of the prior, make sure
it is on the variance scale.}
\usage{
parse_weight_prior(weight_prior = 1, type = "map")
}
\arguments{
\item{weight_prior}{weight of the prior specified on SD scale}

\item{type}{type of fit, e.g. `"map"` or `"pls"`}
}
\description{
Parse the weight of the prior, make sure
it is on the variance scale.
}
