% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.map_estimates.R
\name{plot_eta}
\alias{plot_eta}
\title{Visual representation of eta estimate}
\usage{
plot_eta(x)
}
\arguments{
\item{x}{estimated individual coefficient for parameter (=eta in NONMEM)}
}
\description{
Visual representation of eta estimate
}
