% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.map_estimates.R
\name{print.map_estimates}
\alias{print.map_estimates}
\title{Print MAP estimates}
\usage{
\method{print}{map_estimates}(x, ...)
}
\arguments{
\item{x}{output object from get_map_estimates}

\item{...}{etc}
}
\description{
Print MAP estimates
}
