% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_its.R
\name{run_its}
\alias{run_its}
\title{Crude implementation of Iterative Two State Bayesian (ITSB) estimation method for PopPK data}
\usage{
run_its(parameters = NULL, omega = NULL, err = 100, regimen = NULL,
  model = NULL, max_iter = 5, data = NULL, min_crit = 0.01, ...)
}
\arguments{
\item{parameters}{list of intial parameter estimates}

\item{omega}{omega lower triangle (IIV)}

\item{err}{additive error}

\item{regimen}{`PKPDsim` regimen}

\item{model}{`PKPDsim` model}

\item{max_iter}{maximum number of iterations}

\item{data}{observed data}

\item{min_crit}{minimization criterion (default = 0.01)}

\item{...}{additional arguments passed to `get_map_estimates` function}
}
\description{
Note: this estimation method is mostly for demonstration purposes
as it is unstable, biased, and imprecise. The implementation
in this package is also pretty slow. It is however
a nice example estimation method e.g. for teaching the concept
of estimation methods in pharmacokinetics.
}
