% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sequential_map.R
\name{run_sequential_map}
\alias{run_sequential_map}
\title{Perform a sequential MAP Bayesian fit.}
\usage{
run_sequential_map(model = NULL, regimen = NULL, parameters = NULL,
  covariates = NULL, omega = NULL, ruv = NULL, data = NULL,
  breaks_by_data = TRUE, breaks_by_time = FALSE, break_interval = 24,
  breaks = NULL, weights_prior = NULL, A_init = NULL, weight_prior = 1,
  weight_focus = 1, weight_nonfocus = 0, verbose = FALSE,
  progress = TRUE, ...)
}
\arguments{
\item{model}{PKPDsim model}

\item{regimen}{PKPDsim regimen}

\item{parameters}{population values for parameters for `model``}

\item{covariates}{covariates for `model`, if present}

\item{omega}{vector or matrix specifying lower triangle or full variance-covariance matrix for between-individual variability}

\item{ruv}{list specifying residual variability model (see `PKPDsim` for more details)}

\item{data}{data.frame with `t` and `y` columns}

\item{breaks_by_data}{set breaks automatically (`T`/`F``), 1 data point per section (end of section).}

\item{breaks_by_time}{set breaks automatically by time (`T`/`F``), also specify `break_interval`}

\item{break_interval}{when `breaks_by_time` is `TRUE`, the interval between breaks}

\item{breaks}{vector of breaks for sections}

\item{A_init}{vector of initial state}

\item{weight_prior}{defaults to 1. Weight of population priors specified through `omega`}

\item{weight_focus}{defaults to 1. Weight for the points in the sequential fit that are in the section of focus.}

\item{weight_nonfocus}{defaults to 0. Weight for the points in the sequential fit that are not in the focused section (but later).}

\item{verbose}{verbose output?}

\item{progress}{show progress bar, default is TRUE}

\item{...}{additional arguments passed to `get_map_estimates()` and `sim_ode()`.}
}
\description{
This will allow "tracking" parameters over time, i.e. a form of between-occasion variability
}

