% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_to_full.R
\name{triangle_to_full}
\alias{triangle_to_full}
\title{Convert lower/upper triangle of a matrix to full matrix}
\usage{
triangle_to_full(vect)
}
\arguments{
\item{vect}{vector describing the lower or upper triangle of a matrix}
}
\description{
Convert lower/upper triangle of a matrix to full matrix
}

