% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_by_time.R
\name{weight_by_time}
\alias{weight_by_time}
\title{Gradient weighting by time}
\usage{
weight_by_time(
  time = NULL,
  min_weight = 0,
  t_end_gradient = NULL,
  t_start_gradient = NULL
)
}
\arguments{
\item{time}{vector of time}

\item{min_weight}{minimum weight, default is 0.}

\item{t_end_gradient}{time after which weight should be 1. Default is 0
(all weights are 1)}

\item{t_start_gradient}{time after which gradient should increase from 
`min_weight`. Default is 0.}
}
\description{
Weighting increases from 0 to 1 by a linear gradient. At `t_end_gradient` 
the weighting is 1.
}
