% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_res_from_prob.R
\name{calc_res_from_prob}
\alias{calc_res_from_prob}
\title{Calculate residuals from probabilities for datapoints}
\usage{
calc_res_from_prob(p)
}
\arguments{
\item{p}{probability, e.g. from fit object likelihood info}
}
\description{
calculates the residual associated with a certain probability
It's basically the inverse of what we do in the likelihood function. There
we calculate the probability of a certain point, by calculating
}
\details{
dnorm(obs - ipred, mean, sd) / dnorm(0, mean, sd).

Here, we want to do the reverse, i.e. we have a probability but want to get
the residual. This is useful for censored observations, where we don't 
actually have the "observed", but we can still calculate a residual-
equivalent using this function.
}
