% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_residuals.R
\name{calc_residuals}
\alias{calc_residuals}
\title{Function to calculate residuals from a fit,
based on a fitted parameter set, and data object.}
\usage{
calc_residuals(
  obj,
  data,
  model,
  parameters_population,
  covariates,
  regimen,
  lagtime,
  omega_full,
  error,
  weights,
  transf,
  A_init_population,
  A_init_individual,
  t_init = 0,
  iov_bins = NULL,
  output_include = c(),
  int_step_size = 0.01,
  censoring = NULL,
  censoring_idx = NULL,
  data_before_init = NULL,
  ltbs = FALSE,
  nonfixed = NULL,
  as_eta = c(),
  steady_state_analytic = NULL,
  ...
)
}
\arguments{
\item{obj}{temporary storage object in `get_map_estimates()`}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{parameters_population}{parameters for population predictions}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{regimen}{regimen}

\item{lagtime}{vector of lagtimes for each compartment in the model, either
numeric vector, or vector of parameters.}

\item{omega_full}{full omega matrix}

\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Used in both MAP and NP methods. Note that `weights` argument will also 
affect residuals (residuals will be scaled too).}

\item{transf}{transformation function}

\item{A_init_population}{init vector for model state (population)}

\item{A_init_individual}{init vector for model state (individual)}

\item{t_init}{initialization time before first dose, default 0.}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to 
PKPDsim.}

\item{output_include}{passed to PKPDsim::sim_ode(), returns covariates and
parameter values over time in return object. Only invoked if `residuals` 
option is `TRUE`.}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{censoring}{label for column specifying censoring. If value in dataset 
in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{censoring_idx}{vector with indices for censoring}

\item{data_before_init}{data.frame with data before initial dose}

\item{ltbs}{log-transform both sides? (`NULL` by default, meaning that it 
will be picked up from the PKPDsim model. Can be overridden with `TRUE`). 
Note: `error` should commonly only have additive part.}

\item{nonfixed}{character vector of non-fixed parameter names}

\item{as_eta}{character vector of parameters estimated directly as eta}

\item{steady_state_analytic}{steady state settings (or NULL)}

\item{...}{parameters passed on to `sim_ode()` function}
}
\description{
Function to calculate residuals from a fit,
based on a fitted parameter set, and data object.
}
