% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Some basic input checking}
\usage{
check_inputs(model, data, parameters, omega, regimen, censoring, type)
}
\arguments{
\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{parameters}{list of parameters}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{regimen}{regimen}

\item{censoring}{label for column specifying censoring. If value in dataset 
in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{type}{estimation type, options are `map`, `ls`, and `np_hybrid`}
}
\description{
Some basic input checking
}
