% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid_around_parameters.R
\name{create_grid_around_parameters}
\alias{create_grid_around_parameters}
\title{Create an expanded multivariate grid of parameters around a given set of parameters}
\usage{
create_grid_around_parameters(parameters = list(), grid_size = 4,
  span = 0.2)
}
\arguments{
\item{parameters}{list of parameters}

\item{grid_size}{number of points per parameters}

\item{span}{width of the grid in either side}
}
\description{
The expanded grid outputted from this function can be used for performing 
extended grid non-parametric estimation
}

