% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iov_object.R
\name{create_iov_object}
\alias{create_iov_object}
\title{Calculate the specs for including IOV, e.g. which elements should be fixed
and the right size of the new omega matrix.
Currently only working for a single parameter!}
\usage{
create_iov_object(cv = list(CL = 0.1), omega = c(0.1), data = NULL,
  bins = c(0, 24, 48, 9999), parameters = list(CL = 5), fixed = NULL,
  n = NULL, verbose = TRUE)
}
\arguments{
\item{cv}{list of CVs of parameters}

\item{omega}{omega matrix (lower triangle)}

\item{data}{dataset}

\item{bins}{vector of bins for IOV}

\item{parameters}{named list of parameter values}

\item{fixed}{vector of fixed parameters}

\item{n}{number of IOV elements, will be determined automatically from data, `n` will override if not NULL.}

\item{verbose}{verbosity (`TRUE` or `FALSE`)}

\item{reverse_time}{reverse the timing of the bins to mean "t before last observation". `FALSE` by default.}
}
\description{
Calculate the specs for including IOV, e.g. which elements should be fixed
and the right size of the new omega matrix.
Currently only working for a single parameter!
}
