% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_individual_parameters_from_fit.R
\name{get_individual_parameters_from_fit}
\alias{get_individual_parameters_from_fit}
\title{Get the individual parameters from the fitted coefficients (etas)}
\usage{
get_individual_parameters_from_fit(
  fit,
  parameters,
  nonfixed = NULL,
  as_eta = NULL
)
}
\arguments{
\item{fit}{fit object}

\item{parameters}{list of population parameters}
}
\description{
Get the individual parameters from the fitted coefficients (etas)
}
