% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_estimates.R
\name{get_map_estimates}
\alias{get_map_estimates}
\title{Get MAP estimates}
\usage{
get_map_estimates(
  model = NULL,
  data = NULL,
  parameters = NULL,
  covariates = NULL,
  fixed = c(),
  as_eta = c(),
  omega = NULL,
  weight_prior = 1,
  iov_bins = NULL,
  error = NULL,
  ltbs = NULL,
  obs_type_label = NULL,
  censoring = NULL,
  weights = NULL,
  steady_state_analytic = NULL,
  include_omega = TRUE,
  include_error = TRUE,
  regimen = NULL,
  lagtime = NULL,
  t_init = 0,
  int_step_size = 0.1,
  ll_func = ll_func_PKPDsim,
  optimizer = "optim",
  method = "BFGS",
  control = list(reltol = 1e-05),
  allow_obs_before_dose = FALSE,
  type = "map",
  np_settings = list(),
  cols = list(x = "t", y = "y"),
  residuals = TRUE,
  verbose = FALSE,
  A_init = NULL,
  skip_hessian = FALSE,
  output_include = list(covariates = FALSE, parameters = FALSE)
)
}
\arguments{
\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{parameters}{list of parameters}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{as_eta}{vector of parameters that are estimates as eta (e.g. IOV)}

\item{omega}{between subject variability, supplied as vector specifiying the
lower triangle of the covariance matrix of random effects}

\item{weight_prior}{weighting of priors in relationship to observed data, 
default = 1}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to 
PKPDsim.}

\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}

\item{ltbs}{log-transform both sides? (`NULL` by default, meaning that it 
will be picked up from the PKPDsim model. Can be overridden with `TRUE`). 
Note: `error` should commonly only have additive part.}

\item{obs_type_label}{column name in `data` referring to observation type. 
Can be used for specification of different residual error models for 
differing observation types (e.g. venous and capillary or parent and 
metabolite), Residual error should then be specified as list of vectors, 
e.g. `list(prop = c(0.2, 0.1), add = c(1, 2))`.}

\item{censoring}{label for column specifying censoring. If value in dataset 
in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Note that `weights` argument will also affect residuals (residuals will be 
scaled too).}

\item{steady_state_analytic}{list object with settings for steady state MAP 
estimation.}

\item{include_omega}{TRUE}

\item{include_error}{TRUE}

\item{regimen}{regimen}

\item{lagtime}{vector of lagtimes for each compartment in the model, either
numeric vector, or vector of parameters.}

\item{t_init}{initialization time before first dose, default 0.}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{ll_func}{likelihood function, default is `ll_func_PKPDsim` as included 
in this package.}

\item{optimizer}{optimization library to use, default is `optim`}

\item{method}{optimization method, default `BFGS`. This method is a
gradient-based method, in which the gradients are computed using finite-
difference method. When the model contains a step-function where a
estimated parameter is involved in the step-function (such as in the case of
lagtime), it is advised to use a non-gradient-based estimation method
such as `Nelder-Mead` to avoid estimation failures.}

\item{control}{list of options passed to `optim()` function}

\item{allow_obs_before_dose}{allow observation before first dose?}

\item{type}{estimation type, options are `map`, `ls`, and `np_hybrid`}

\item{np_settings}{list with settings for non-parametric estimation (if 
selected), containing any of the following: `error`, `grid_span`, grid_size`,
 `grid_exponential`}

\item{cols}{column names}

\item{residuals}{show residuals? This requires an additional simulation so 
will be slightly slower.}

\item{verbose}{show more output}

\item{A_init}{initial state vector}

\item{skip_hessian}{skip calculation of Hessian}

\item{output_include}{passed to PKPDsim::sim_ode(), returns covariates and
parameter values over time in return object. Only invoked if `residuals` 
option is `TRUE`.}
}
\description{
Get MAP estimates
}
\examples{
\dontrun{
## define parameters
pk1 <- new_ode_model(
  code = "dAdt[1] = -(CL/V)*A[1]", 
  obs = list(scale = "V", cmt=1)
)
regimen <- new_regimen(
  amt = 100, interval = 12, times=c(0,24), 
  type="infusion"
)
parameters <- list("CL" = 3, "V" = 10) 
omega <- c(0.1, 
           0.02, 0.1)

## simulate single individual
data <- sim_ode(
  ode = pk1, 
  parameters = list(CL = 6, V = 14), 
  regimen = regimen,
  omega = omega,
  t_obs = c(23, 25, 27, 32, 36)
)

## Fit
get_map_estimates(
  model = pk1, 
  data = data, 
  parameters = parameters,
  omega = omega,
  regimen = regimen,
  error = list(prop = 0.1, add = 0.1)
)
}

}
