% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_estimates.R
\name{get_map_estimates}
\alias{get_map_estimates}
\title{Get MAP estimates}
\usage{
get_map_estimates(model = NULL, data = NULL, parameters = NULL,
  covariates = NULL, fixed = c(), as_eta = c(), weights = NULL,
  omega = NULL, w_prior = 1, error = list(prop = 0.1, add = 0.1, exp = 0),
  include_omega = TRUE, include_error = TRUE, regimen = NULL,
  int_step_size = 0.1, method = "L-BFGS-B", type = "map", cols = list(x
  = "t", y = "y"), residuals = TRUE, verbose = FALSE, checks = FALSE, ...)
}
\arguments{
\item{model}{model}

\item{data}{data}

\item{parameters}{parameters}

\item{covariates}{covariates}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{as_eta}{vector of parameters that are estimates as eta (e.g. IOV)}

\item{weights}{vector of weights. Length of vector should be same as length of observation vector. If NULL, all weights are 1.}

\item{omega}{between subject variability, supplied as vector specifiying the lower triangle of the covariance matrix of random effects}

\item{w_prior}{weighting of priors in relationship to observed data, default = 1}

\item{error}{residual error, specified as list with arguments `add` and/or `prop` specifying the additive and proportional parts}

\item{include_omega}{TRUE}

\item{include_error}{TRUE}

\item{regimen}{regimen}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{method}{optimization method, default L-BFGS-B}

\item{type}{estimation type, currently only option is `map`}

\item{cols}{column names}

\item{residuals}{show residuals? This requires an additional simulation so will be slightly slower.}

\item{verbose}{show more output}

\item{checks}{perform input checks}

\item{...}{parameters passed on to `sim_ode()` function}
}
\description{
Get MAP estimates
}

