% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_estimates.R
\name{get_map_estimates}
\alias{get_map_estimates}
\title{Get MAP estimates}
\usage{
get_map_estimates(model = NULL, data = NULL, parameters = NULL,
  covariates = NULL, fixed = c(), as_eta = c(), omega = NULL,
  weight_prior = 1, iov_bins = NULL, error = NULL, ltbs = NULL,
  censoring = NULL, mixture = NULL, weights = NULL,
  include_omega = TRUE, include_error = TRUE, regimen = NULL,
  t_init = 0, int_step_size = 0.1, optimizer = "optim", method = "BFGS",
  control = list(reltol = 1e-05), allow_obs_before_dose = FALSE,
  type = "map", np_settings = list(), cols = list(x = "t", y = "y"),
  residuals = TRUE, verbose = FALSE, A_init = NULL,
  skip_hessian = FALSE, output_include = list(covariates = FALSE, parameters
  = FALSE), ...)
}
\arguments{
\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{parameters}{list of parameters}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{as_eta}{vector of parameters that are estimates as eta (e.g. IOV)}

\item{omega}{between subject variability, supplied as vector specifiying the lower triangle of the covariance matrix of random effects}

\item{weight_prior}{weighting of priors in relationship to observed data, default = 1}

\item{iov_bins}{bins for inter-occasion variability. Passed unchanged to PKPDsim.}

\item{error}{residual error, specified as list with arguments `add` and/or `prop` specifying the additive and proportional parts}

\item{ltbs}{log-transform both sides? (`NULL` by default, meaning that it will be picked up from the PKPDsim model. Can be overridden with `TRUE`). Note: `error` should commonly only have additive part.}

\item{censoring}{label for column specifying censoring. If value in dataset in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{mixture}{specify mixture model. Currently for single parameter only. Overwrites regular parameter, if specified. Specify e.g. as: `mixture = list("CL" = c(5, 9))`}

\item{weights}{vector of weights for error. Length of vector should be same as length of observation vector. If NULL (default), all weights are equal. Used in both MAP and NP methods. Note that `weights` argument will also affect residuals (residuals will be scaled too).}

\item{include_omega}{TRUE}

\item{include_error}{TRUE}

\item{regimen}{regimen}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{optimizer}{optimization library to use, default is `optim`}

\item{method}{optimization method, default `BFGS`}

\item{control}{list of options passed to `optim()` function}

\item{allow_obs_before_dose}{allow observation before first dose?}

\item{type}{estimation type, options are `map`, `ls`, and `np_hybrid`}

\item{np_settings}{list with settings for non-parametric estimation (if selected), containing any of the following: `error`, `grid_span`, grid_size`, `grid_exponential`}

\item{cols}{column names}

\item{residuals}{show residuals? This requires an additional simulation so will be slightly slower.}

\item{verbose}{show more output}

\item{A_init}{initial state vector}

\item{skip_hessian}{skip calculation of Hessian (in `bbmle::mle2()`)}

\item{output_include}{passed to PKPDsim::sim_ode(), returns covariates and parmeter values over time in return object. Only invoked if `residuals` option is `TRUE`.}

\item{...}{parameters passed on to `sim_ode()` function}
}
\description{
Get MAP estimates
}
