% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_map_estimates.R
\name{get_map_estimates}
\alias{get_map_estimates}
\title{Get MAP estimates}
\usage{
get_map_estimates(model = NULL, data = NULL, parameters = NULL,
  covariates = NULL, fixed = NULL, weights = NULL, omega = NULL,
  error = list(prop = 0.1, add = 0.1, exp = 0), regimen = NULL,
  int_step_size = 0.1, method = "L-BFGS-B", cols = list(x = "t", y = "y"),
  verbose = FALSE)
}
\arguments{
\item{model}{model}

\item{data}{data}

\item{parameters}{parameters}

\item{covariates}{covariates}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{weights}{vector of weights. Length of vector should be same as length of observation vector. If NULL, all weights are 1.}

\item{omega}{between subject variability, supplied as vector specifiying the lower triangle of the covariance matrix of random effects}

\item{error}{residual error, specified as list with arguments `add` and/or `prop` specifying the additive and proportional parts}

\item{regimen}{regimen}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{method}{optimization method, default L-BFGS-B}

\item{cols}{column names}

\item{verbose}{show more output}
}
\description{
Get MAP estimates
}

