% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_estimates.R
\name{get_map_estimates}
\alias{get_map_estimates}
\title{Get MAP estimates}
\usage{
get_map_estimates(model = NULL, data = NULL, parameters = NULL,
  covariates = NULL, fixed = c(), as_eta = c(), omega = NULL,
  weight_prior = 1, error = list(prop = 0.1, add = 0.1, exp = 0),
  weights = NULL, include_omega = TRUE, include_error = TRUE,
  regimen = NULL, int_step_size = 0.1, optimizer = "optim",
  method = "BFGS", control = list(reltol = 1e-05), type = "map",
  np_settings = list(), cols = list(x = "t", y = "y"), residuals = TRUE,
  verbose = FALSE, A_init = NULL, skip_hessian = FALSE,
  output_include = list(covariates = FALSE, parameters = FALSE), ...)
}
\arguments{
\item{model}{model, created using `PKPDsim::new_ode_model()`}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{parameters}{list of parameters}

\item{covariates}{list of covariates, each one created using `PKPDsim::new_coviarate()`}

\item{fixed}{fix a specific parameters, supplied as vector of strings}

\item{as_eta}{vector of parameters that are estimates as eta (e.g. IOV)}

\item{omega}{between subject variability, supplied as vector specifiying the lower triangle of the covariance matrix of random effects}

\item{weight_prior}{weighting of priors in relationship to observed data, default = 1}

\item{error}{residual error, specified as list with arguments `add` and/or `prop` specifying the additive and proportional parts}

\item{weights}{vector of weights for error. Length of vector should be same as length of observation vector. If NULL (default), all weights are equal. Used in both MAP and NP methods. Note that `weights` argument will also affect residuals (residuals will be scaled too).}

\item{include_omega}{TRUE}

\item{include_error}{TRUE}

\item{regimen}{regimen}

\item{int_step_size}{integrator step size passed to PKPDsim}

\item{optimizer}{optimization library to use, default is `optim`}

\item{method}{optimization method, default `L-BFGS-B`}

\item{control}{list of options passed to `optim()` function}

\item{type}{estimation type, options are `map`, `ls`, and `np_hybrid`}

\item{np_settings}{list with settings for non-parametric estimation (if selected), containing any of the following: `error`, `grid_span`, grid_size`, `grid_exponential`}

\item{cols}{column names}

\item{residuals}{show residuals? This requires an additional simulation so will be slightly slower.}

\item{verbose}{show more output}

\item{A_init}{initial state vector}

\item{skip_hessian}{skip calculation of Hessian (in `bbmle::mle2()`)}

\item{output_include}{passed to PKPDsim::sim_ode(), returns covariates and parmeter values over time in return object. Only invoked if `residuals` option is `TRUE`.}

\item{...}{parameters passed on to `sim_ode()` function}
}
\description{
Get MAP estimates
}
