% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npar_estimates.R
\name{get_npar_estimates}
\alias{get_npar_estimates}
\title{Get non-parametric estimates given a set of support points, model, and data}
\usage{
get_npar_estimates(parameter_grid = NULL, error = list(prop = 0.05, add =
  0.1), model = NULL, covariates = NULL, regimen = NULL, t_obs = c(24),
  data = NULL, ...)
}
\arguments{
\item{error}{error model}

\item{model}{`PKPDsim` model}

\item{regimen}{`PKPDsim` regimen}

\item{t_obs}{vector of observations}

\item{data}{vector of obsersved data}

\item{parameters}{grid (data.frame) of support points}
}
\description{
Get non-parametric estimates given a set of support points, model, and data
}

