% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varcov_matrix.R
\name{get_varcov_matrix}
\alias{get_varcov_matrix}
\title{Performs some safety checks on the varcov matrix and returns fallback if 
not positive definite matrix or if NULL.}
\usage{
get_varcov_matrix(vcov, fallback = NULL)
}
\arguments{
\item{vcov}{variance covariance matrix (e.g. from fit)}

\item{fallback}{a fallback variance-covariance matrix, if the varcov-matrix
from the fit object does not satisfy safety checks (e.g. not postive-
definite).}
}
\description{
Performs some safety checks on the varcov matrix and returns fallback if 
not positive definite matrix or if NULL.
}
