% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_func_PKPDsim.R
\name{ll_func_PKPDsim}
\alias{ll_func_PKPDsim}
\title{Likelihood function for MAP optimization using PKPDsim}
\usage{
ll_func_PKPDsim(
  data,
  sim_object,
  parameters,
  covariates = covariates,
  nonfixed,
  eta01,
  eta02,
  eta03,
  eta04,
  eta05,
  eta06,
  eta07,
  eta08,
  eta09,
  eta10,
  eta11,
  eta12,
  eta13,
  eta14,
  eta15,
  eta16,
  eta17,
  eta18,
  eta19,
  eta20,
  eta21,
  eta22,
  eta23,
  eta24,
  error = error,
  transf,
  model,
  omega_full,
  omega_inv,
  omega_eigen,
  sig,
  weights,
  as_eta,
  censoring_idx,
  censoring_label,
  iov_bins,
  t_init = 0,
  calc_ofv,
  regimen,
  lagtime = c(0),
  steady_state_analytic,
  include_omega,
  include_error,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{vector of}

\item{sim_object}{design (event-table) obtained from PKPDsim to be used in simulations/optimizations}

\item{parameters}{parameter list}

\item{covariates}{covariates list}

\item{nonfixed}{non-fixed (i.e. estimated) parameters}

\item{eta01}{eta1}

\item{eta02}{eta2}

\item{eta03}{eta3}

\item{eta04}{eta4}

\item{eta05}{eta5}

\item{eta06}{eta6}

\item{eta07}{eta7}

\item{eta08}{eta8}

\item{eta09}{eta9}

\item{eta10}{eta10}

\item{eta11}{eta11}

\item{eta12}{eta12}

\item{eta13}{eta13}

\item{eta14}{eta14}

\item{eta15}{eta15}

\item{eta16}{eta16}

\item{eta17}{eta17}

\item{eta18}{eta19}

\item{eta19}{eta19}

\item{eta20}{eta20}

\item{eta21}{eta21}

\item{eta22}{eta22}

\item{eta23}{eta23}

\item{eta24}{eta24}

\item{error}{error model to use, e.g. `list(add = .5, prop = .15)`}

\item{transf}{transformation function for data, if needed. E.g. `log(x)`. Default is `function(x) = x`.}

\item{model}{PKPDsim model}

\item{omega_full}{full omega matrix}

\item{omega_inv}{inverse of omega matrix. Passed to this function to avoid doing computations in each iteration of the search.}

\item{omega_eigen}{eigenvalue decomposation (logged) of omega matrix. Passed to this function to avoid doing computations in each iteration of the search.}

\item{sig}{signficance, used in optimization}

\item{weights}{weights for data, generally a vector of weights between 0 and 1.}

\item{as_eta}{implement as regular eta instead of exponential eta, can be vector.}

\item{censoring_idx}{censoring indices, used for <LOQ data.}

\item{censoring_label}{censoring label, used for <LOQ data}

\item{iov_bins}{IOV bins object}

\item{t_init}{init time for simulations, default 0.}

\item{calc_ofv}{function to calculate OFV based on simulated data and set of parameters and omega matrix}

\item{regimen}{PKPDsim regimen}

\item{steady_state_analytic}{list object with settings for steady state MAP estimation.}

\item{include_omega}{include omega in calculation of OFV?}

\item{include_error}{include residual error in calculation of OFV?}

\item{verbose}{verbose output?}

\item{...}{passed on to PKPDsim}
}
\description{
Likelihood function for MAP optimization using PKPDsim
}
