% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_input_data.R
\name{parse_input_data}
\alias{parse_input_data}
\title{Parse input data for fit}
\usage{
parse_input_data(data)
}
\arguments{
\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}
}
\description{
Parse input data for fit
}
