% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_residuals_from_predictions.R
\name{parse_residuals_from_predictions}
\alias{parse_residuals_from_predictions}
\title{Parse simulated data and extract residuals}
\usage{
parse_residuals_from_predictions(
  obj,
  sim_ipred,
  sim_pred,
  data,
  omega_full,
  transf,
  error,
  censoring,
  censoring_idx,
  data_before_init,
  weights
)
}
\arguments{
\item{obj}{temporary storage object in `get_map_estimates()`}

\item{sim_ipred}{data frame with simulated individual predictions}

\item{sim_pred}{data frame with simulated population predictions}

\item{data}{data data.frame with columns `t` and `y` (and possibly evid)}

\item{omega_full}{full omega matrix}

\item{transf}{transformation function}

\item{error}{residual error, specified as list with arguments `add` and/or 
`prop` specifying the additive and proportional parts}

\item{censoring}{label for column specifying censoring. If value in dataset 
in this column is < 0 then censoring is assumed <LLOQ. If > 0 then  >ULOQ.}

\item{censoring_idx}{vector with indices for censoring}

\item{data_before_init}{data.frame with data before initial dose}

\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Note that `weights` argument will also affect residuals (residuals will be 
scaled too).}
}
\description{
Parse simulated data and extract residuals
}
