% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_weights.R
\name{parse_weights}
\alias{parse_weights}
\title{Parse weights argument}
\usage{
parse_weights(weights, t_obs)
}
\arguments{
\item{weights}{vector of weights for error. Length of vector should be same 
as length of observation vector. If NULL (default), all weights are equal. 
Used in both MAP and NP methods. Note that `weights` argument will also 
affect residuals (residuals will be scaled too).}
}
\description{
Parse weights argument
}
