% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sequential_map.R
\name{run_sequential_map}
\alias{run_sequential_map}
\title{Perform a sequential MAP Bayesian fit.}
\usage{
run_sequential_map(model = NULL, regimen = NULL, parameters = list(),
  covariates = NULL, data = NULL, breaks_by_data = TRUE,
  breaks_by_time = FALSE, break_interval = 24, breaks = NULL,
  weights_prior = NULL, A_init = NULL, weight_prior = 1,
  weight_unfocus = 0.1, ...)
}
\arguments{
\item{model}{PKPDsim model}

\item{regimen}{PKPDsim regimen}

\item{parameters}{population values for parameters for `model``}

\item{covariates}{covariates for `model`, if present}

\item{data}{data.frame with `t` and `y` columns}

\item{breaks_by_data}{set breaks automatically (`T`/`F``), 1 data point per section (end of section).}

\item{breaks_by_time}{set breaks automatically by time (`T`/`F``), also specify `break_interval`}

\item{break_interval}{when `breaks_by_time` is `TRUE`, the interval between breaks}

\item{breaks}{vector of breaks for sections}

\item{A_init}{vector of initial state}

\item{weight_prior}{weight of population priors specified through `omega`}

\item{weight_unfocus}{weight for the points in the sequential fit that are not in the focused section (but later).}

\item{...}{additional arguments passed to `get_map_estimates()` and `sim_ode()`.}
}
\description{
This will allow "tracking" parameters over time, i.e. a form of between-occasion variability
}

