% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cwres.R
\name{simulate_with_etas}
\alias{simulate_with_etas}
\title{Simulate predictions with a given eta vector}
\usage{
simulate_with_etas(
  eta,
  parameters_population,
  nonfixed,
  as_eta,
  model,
  covariates,
  regimen,
  lagtime,
  t_obs,
  obs_type,
  int_step_size,
  iov_bins,
  A_init,
  t_init,
  steady_state_analytic
)
}
\description{
Helper for computing the Jacobian in CWRES calculation via finite
differences. Computes individual parameters from etas and runs a
PKPDsim simulation.
}
\keyword{internal}
